# src/main/java/com/postech/workshop_service/api/controllers/openapi.yaml
openapi: 3.0.3
info:
  title: Workshop Service API
  description: API REST para gerenciamento de clientes e veiculos da oficina.
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local

security: []

paths:
  /api/v1/clientes:
    post:
      tags: [Clientes]
      summary: Criar um novo cliente
      operationId: criarCliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CadastroClienteRequest'
      responses:
        '201':
          description: Cliente criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClienteResponse'
        '422':
          description: Erro de validacao estrutural no payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags: [Clientes]
      summary: Listar clientes com paginacao
      operationId: listarClientes
      parameters:
        - name: pagina
          in: query
          schema:
            type: integer
            default: 0
        - name: tamanho
          in: query
          schema:
            type: integer
            default: 20
        - name: incluirInativos
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClienteResponse'

  /api/v1/clientes/{id}:
    get:
      tags: [Clientes]
      summary: Buscar cliente por ID
      operationId: buscarClientePorId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: incluirInativos
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Cliente encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClienteResponse'
        '404':
          description: Cliente nao encontrado

    put:
      tags: [Clientes]
      summary: Atualizar dados de um cliente
      operationId: atualizarCliente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizarClienteRequest'
      responses:
        '200':
          description: Cliente atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClienteResponse'
        '404':
          description: Cliente nao encontrado
        '422':
          description: Erro de validacao estrutural no payload

    delete:
      tags: [Clientes]
      summary: Remover um cliente (soft delete)
      operationId: removerCliente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Cliente removido logicamente

  /api/v1/clientes/documento/{documento}:
    get:
      tags: [Clientes]
      summary: Buscar cliente por CPF ou CNPJ
      operationId: buscarClientePorDocumento
      parameters:
        - name: documento
          in: path
          required: true
          schema:
            type: string
        - name: incluirInativos
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Cliente encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClienteResponse'
        '404':
          description: Cliente nao encontrado

  /api/v1/veiculos:
    post:
      tags: [Veiculos]
      summary: Cadastrar veiculo
      operationId: criarVeiculo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CadastroVeiculoRequest'
      responses:
        '201':
          description: Veiculo criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VeiculoResponse'
        '422':
          description: Erro de validacao estrutural no payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags: [Veiculos]
      summary: Listar veiculos com paginacao e filtros
      operationId: listarVeiculos
      parameters:
        - name: pagina
          in: query
          schema:
            type: integer
            default: 0
        - name: tamanho
          in: query
          schema:
            type: integer
            default: 20
        - name: placa
          in: query
          schema:
            type: string
        - name: clienteId
          in: query
          schema:
            type: string
            format: uuid
        - name: incluirInativos
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Pagina de veiculos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginaVeiculosResponse'

  /api/v1/veiculos/{id}:
    get:
      tags: [Veiculos]
      summary: Buscar veiculo por identificador
      operationId: buscarVeiculoPorId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: incluirInativos
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Veiculo encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VeiculoResponse'
        '404':
          description: Veiculo nao encontrado

    put:
      tags: [Veiculos]
      summary: Atualizar veiculo
      operationId: atualizarVeiculo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizarVeiculoRequest'
      responses:
        '200':
          description: Veiculo atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VeiculoResponse'
        '404':
          description: Veiculo nao encontrado
        '422':
          description: Erro de validacao estrutural no payload

    delete:
      tags: [Veiculos]
      summary: Remover veiculo logicamente
      operationId: removerVeiculo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Veiculo removido logicamente

  /api/v1/veiculos/placa/{placa}:
    get:
      tags: [Veiculos]
      summary: Buscar veiculo por placa
      operationId: buscarVeiculoPorPlaca
      parameters:
        - name: placa
          in: path
          required: true
          schema:
            type: string
        - name: incluirInativos
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Veiculo encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VeiculoResponse'
        '404':
          description: Veiculo nao encontrado

  /api/v1/veiculos/cliente/{clienteId}:
    get:
      tags: [Veiculos]
      summary: Listar veiculos por cliente
      operationId: listarVeiculosPorCliente
      parameters:
        - name: clienteId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: incluirInativos
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Lista de veiculos do cliente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VeiculoResponse'

  /api/v1/veiculos/{id}/clientes/{clienteId}:
    post:
      tags: [Veiculos]
      summary: Vincular cliente ao veiculo
      operationId: vincularClienteVeiculo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: clienteId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Cliente vinculado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VeiculoResponse'
        '404':
          description: Veiculo ou cliente nao encontrado
        '400':
          description: Regra de negocio violada

    delete:
      tags: [Veiculos]
      summary: Desvincular cliente do veiculo
      operationId: desvincularClienteVeiculo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: clienteId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Cliente desvinculado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VeiculoResponse'
        '404':
          description: Veiculo ou cliente nao encontrado
        '400':
          description: Regra de negocio violada

components:
  schemas:
    CadastroClienteRequest:
      type: object
      required: [nome, documento, email]
      properties:
        nome:
          type: string
          example: 'Joao da Silva'
        documento:
          type: string
          description: 'CPF (11 digitos) ou CNPJ (14 digitos)'
          example: '98765432100'
        email:
          type: string
          example: 'joao@email.com'
        telefone:
          type: string
          example: '11999999999'
      example:
        nome: 'Joao da Silva'
        documento: '98765432100'
        email: 'joao@email.com'
        telefone: '11999999999'

    AtualizarClienteRequest:
      type: object
      properties:
        nome:
          type: string
          example: 'Joao da Silva Junior'
        email:
          type: string
          example: 'joao.junior@email.com'
        telefone:
          type: string
          example: '11888888888'
        endereco:
          $ref: '#/components/schemas/EnderecoDTO'
        dataNascimentoFundacao:
          type: string
          format: date
          example: '1990-01-15'
        observacoes:
          type: string
          example: 'Cliente VIP'
      example:
        nome: 'Joao da Silva Junior'
        email: 'joao.junior@email.com'
        telefone: '11888888888'
        endereco:
          logradouro: 'Rua das Flores'
          numero: '123'
          complemento: 'Apto 4'
          bairro: 'Centro'
          cidade: 'Sao Paulo'
          estado: 'SP'
          cep: '01234567'
        dataNascimentoFundacao: '1990-01-15'
        observacoes: 'Cliente VIP'

    EnderecoDTO:
      type: object
      properties:
        logradouro:
          type: string
          example: 'Rua das Flores'
        numero:
          type: string
          example: '123'
        complemento:
          type: string
          example: 'Apto 4'
        bairro:
          type: string
          example: 'Centro'
        cidade:
          type: string
          example: 'Sao Paulo'
        estado:
          type: string
          example: 'SP'
        cep:
          type: string
          example: '01234567'
      example:
        logradouro: 'Rua das Flores'
        numero: '123'
        complemento: 'Apto 4'
        bairro: 'Centro'
        cidade: 'Sao Paulo'
        estado: 'SP'
        cep: '01234567'

    ClienteResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: '550e8400-e29b-41d4-a716-446655440000'
        nome:
          type: string
          example: 'Joao da Silva'
        documento:
          type: string
          description: 'Documento mascarado (ex: ***.654.321-**)'
          example: '***.654.321-**'
        email:
          type: string
          example: 'joao@email.com'
        telefone:
          type: string
          example: '11999999999'
        endereco:
          $ref: '#/components/schemas/EnderecoDTO'
        dataNascimentoFundacao:
          type: string
          format: date
          example: '1990-01-15'
        observacoes:
          type: string
          example: 'Cliente VIP'
        ativo:
          type: boolean
          example: true
        dataCriacao:
          type: string
          format: date-time
          example: '2024-04-24T10:00:00'
        dataUltimaAtualizacao:
          type: string
          format: date-time
          example: '2024-04-24T10:00:00'
        dataRemocao:
          type: string
          format: date-time
          nullable: true
      example:
        id: '550e8400-e29b-41d4-a716-446655440000'
        nome: 'Joao da Silva'
        documento: '***.654.321-**'
        email: 'joao@email.com'
        telefone: '11999999999'
        endereco:
          logradouro: 'Rua das Flores'
          numero: '123'
          complemento: 'Apto 4'
          bairro: 'Centro'
          cidade: 'Sao Paulo'
          estado: 'SP'
          cep: '01234567'
        dataNascimentoFundacao: '1990-01-15'
        observacoes: 'Cliente VIP'
        ativo: true
        dataCriacao: '2024-04-24T10:00:00'
        dataUltimaAtualizacao: '2024-04-24T10:00:00'
        dataRemocao: null

    CadastroVeiculoRequest:
      type: object
      required: [placa, marca, modelo, ano, clientesIds]
      properties:
        placa:
          type: string
          example: 'ABC1234'
        marca:
          type: string
          example: 'Toyota'
        modelo:
          type: string
          example: 'Corolla'
        ano:
          type: integer
          example: 2024
        cor:
          type: string
          example: 'Prata'
        observacoes:
          type: string
          example: 'Segundo dono'
        clientesIds:
          type: array
          items:
            type: string
            format: uuid
          example:
            - '550e8400-e29b-41d4-a716-446655440000'
      example:
        placa: 'ABC1234'
        marca: 'Toyota'
        modelo: 'Corolla'
        ano: 2024
        cor: 'Prata'
        observacoes: 'Segundo dono'
        clientesIds:
          - '550e8400-e29b-41d4-a716-446655440000'

    AtualizarVeiculoRequest:
      type: object
      properties:
        placa:
          type: string
          example: 'XYZ5678'
        marca:
          type: string
          example: 'Honda'
        modelo:
          type: string
          example: 'Civic'
        ano:
          type: integer
          example: 2023
        cor:
          type: string
          example: 'Preto'
        observacoes:
          type: string
          example: 'Revisao em dia'
      example:
        placa: 'XYZ5678'
        marca: 'Honda'
        modelo: 'Civic'
        ano: 2023
        cor: 'Preto'
        observacoes: 'Revisao em dia'

    VeiculoResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: '660e8400-e29b-41d4-a716-446655440001'
        placa:
          type: string
          example: 'ABC1234'
        marca:
          type: string
          example: 'Toyota'
        modelo:
          type: string
          example: 'Corolla'
        ano:
          type: integer
          example: 2024
        cor:
          type: string
          example: 'Prata'
        observacoes:
          type: string
          example: 'Segundo dono'
        clientes:
          type: array
          items:
            $ref: '#/components/schemas/ClienteVinculadoResponse'
        ativo:
          type: boolean
          example: true
        dataCriacao:
          type: string
          format: date-time
          example: '2024-04-24T10:00:00'
        dataUltimaAtualizacao:
          type: string
          format: date-time
          example: '2024-04-24T10:00:00'
        dataRemocao:
          type: string
          format: date-time
          nullable: true
      example:
        id: '660e8400-e29b-41d4-a716-446655440001'
        placa: 'ABC1234'
        marca: 'Toyota'
        modelo: 'Corolla'
        ano: 2024
        cor: 'Prata'
        observacoes: 'Segundo dono'
        clientes:
          - id: '550e8400-e29b-41d4-a716-446655440000'
            nome: 'Joao da Silva'
            documentoMascarado: '***.654.321-**'
        ativo: true
        dataCriacao: '2024-04-24T10:00:00'
        dataUltimaAtualizacao: '2024-04-24T10:00:00'
        dataRemocao: null

    ClienteVinculadoResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: '550e8400-e29b-41d4-a716-446655440000'
        nome:
          type: string
          example: 'Joao da Silva'
        documentoMascarado:
          type: string
          example: '***.654.321-**'
      example:
        id: '550e8400-e29b-41d4-a716-446655440000'
        nome: 'Joao da Silva'
        documentoMascarado: '***.654.321-**'

    PaginaVeiculosResponse:
      type: object
      properties:
        conteudo:
          type: array
          items:
            $ref: '#/components/schemas/VeiculoResponse'
        pagina:
          type: integer
          example: 0
        tamanho:
          type: integer
          example: 20
        totalElementos:
          type: integer
          example: 1
        totalPaginas:
          type: integer
          example: 1
      example:
        conteudo:
          - id: '660e8400-e29b-41d4-a716-446655440001'
            placa: 'ABC1234'
            marca: 'Toyota'
            modelo: 'Corolla'
            ano: 2024
            cor: 'Prata'
            observacoes: 'Segundo dono'
            clientes:
              - id: '550e8400-e29b-41d4-a716-446655440000'
                nome: 'Joao da Silva'
                documentoMascarado: '***.654.321-**'
            ativo: true
            dataCriacao: '2024-04-24T10:00:00'
            dataUltimaAtualizacao: '2024-04-24T10:00:00'
            dataRemocao: null
        pagina: 0
        tamanho: 20
        totalElementos: 1
        totalPaginas: 1

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: '2024-04-24T10:00:00'
        status:
          type: integer
          example: 422
        error:
          type: string
          example: 'Unprocessable Entity'
        message:
          type: string
          example: 'Erro de validacao estrutural no payload.'
        path:
          type: string
          example: '/api/v1/clientes'
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDetail'
      example:
        timestamp: '2024-04-24T10:00:00'
        status: 422
        error: 'Unprocessable Entity'
        message: 'Erro de validacao estrutural no payload.'
        path: '/api/v1/clientes'
        fieldErrors:
          - field: 'documento'
            message: 'Documento invalido'

    FieldErrorDetail:
      type: object
      properties:
        field:
          type: string
          example: 'documento'
        message:
          type: string
          example: 'Documento invalido'
      example:
        field: 'documento'
        message: 'Documento invalido'